backendpramlearn/requirements.txt
# Core Django & API
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0

# Database & Crypto - GANTI PyMySQL dengan psycopg2
psycopg2-binary==2.9.9
# PyMySQL==1.1.0  # Hapus atau comment ini
cryptography==42.0.5

# File Processing
Pillow==10.4.0
openpyxl==3.1.2

# Data Science & ML (untuk ARCS analysis) - Python 3.12 compatible
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
numpy==1.26.4

# Real-time WebSocket
channels==4.0.0
channels-redis==4.2.0
redis==5.0.7
daphne==4.1.2
uvicorn[standard]==0.24.0  # Tambahkan ini untuk Gunicorn ASGI support

# Production Server
gunicorn==21.2.0

# Development & Testing
Faker==25.0.1
pytest==8.1.1
pytest-django==4.8.0

# Additional utilities
python-slugify==8.0.4
python-decouple==3.8

# Django additional utilities (untuk ARCS motivation profiling)
django-extensions==3.2.3

# Date utilities
python-dateutil==2.9.0

# JSON handling untuk WebSocket
jsonfield==3.1.0

# CORS handling untuk production
whitenoise==6.6.0

# Environment variables
python-dotenv==1.0.1